- Fixed some bugs with float parameters in MEL being interpreted as integer. - Added checkbox for enabling/disabling brush falloff. - Brush tool now implements full use of a falloff and sensitivity parameter.

+++ Stubble/Toolbox/Tools/CutTool/CutTool.cpp
@@ -175,7 +175,7 @@ void CutTool::doCut()
 
 		// where do we cut
 		int segmentsLeft = 0;
-		int segmentCount = guide->mSegmentsAdditionalInfo.size();
+		size_t segmentCount = guide->mSegmentsAdditionalInfo.size();
 		while ( segmentsLeft < segmentCount && ! guide->mSegmentsAdditionalInfo[ segmentsLeft ].mInsideBrush )
 		{
 			segmentsLeft++;
